fix(transfer): rename Heade.tsx to Header.tsx
